# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SignalGen Go API
  version: 0.9.0
  description: |
    Implemented API surface for the incremental SignalGen Go backend.
    The Go API performs email/password and recovery flows through Supabase Auth,
    verifies returned bearer tokens, and issues a separate opaque app-session
    token. Proposed routes that are not implemented are intentionally excluded.
servers:
  - url: http://127.0.0.1:8080
    description: Local Docker development
...
paths:
...
  /api/v1/rules:
    get:
      tags: [Analysis]
      operationId: listRules
      summary: List analysis rules visible to the current account
      description: Returns the frozen system baseline followed by up to 100 rules owned by the current account.
      security:
        - supabaseBearer: []
          appSession: []
      responses:
        "200":
          description: Rule list.
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/EntitlementRequired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      tags: [Analysis]
      operationId: createRule
      summary: Create a rule owned by the current account
      description: A maximum of 100 user rules may exist per account.
      security:
        - supabaseBearer: []
          appSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRuleCreateRequest"
      responses:
        "201":
          description: Owner-scoped rule created.
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRuleResource"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/EntitlementRequired"
        "409":
          description: The account already owns the maximum 100 rules.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/rules/{id}:
    get:
      tags: [Analysis]
      operationId: getRule
      summary: Read a versioned rule snapshot
      security:
        - supabaseBearer: []
          appSession: []
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "200":
          description: System or owner-scoped rule snapshot.
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleResource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/EntitlementRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    patch:
      tags: [Analysis]
      operationId: updateRule
      summary: Update a rule owned by the current account
      description: The submitted version must match the stored version. The system baseline is read-only.
      security:
        - supabaseBearer: []
          appSession: []
      parameters:
        - $ref: "#/components/parameters/RuleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRuleUpdateRequest"
      responses:
        "200":
          description: Updated owner-scoped rule with an incremented version.
          headers:
            Cache-Control:
              $ref: "#/components/headers/PrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRuleResource"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/EntitlementRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Version conflict or attempted mutation of the read-only system rule.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      tags: [Analysis]
      operationId: deleteRule
      summary: Delete a rule owned by the current account
      description: The submitted version must match the stored version. The system baseline is read-only.
      security:
        - supabaseBearer: []
          appSession: []
      parameters:
        - $ref: "#/components/parameters/RuleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRuleDeleteRequest"
      responses:
        "204":
          description: Owner-scoped rule deleted.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/EntitlementRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Version conflict or attempted mutation of the read-only system rule.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/RateLimited"
        "422":
          $ref: "#/components/responses/InvalidRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    supabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued and refreshed by Supabase Auth.
    appSession:
      type: apiKey
      in: header
      name: X-App-Session
      description: Opaque one-time-issued SignalGen app-session secret.
  headers:
    PrivateNoStore:
      description: Protected responses must not be stored by shared or browser caches.
      schema:
        type: string
        example: private, no-store
  parameters:
...
    RuleId:
      name: id
      in: path
      required: true
      schema: {type: string, example: default-scalping-v1}
...
  responses:
    InvalidRequest:
      description: Request syntax, values, or supported capability is invalid.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Bearer or app-session credential is missing or invalid.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    EntitlementRequired:
      description: Required feature grant is missing, expired, or revoked.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: {code: ENTITLEMENT_REQUIRED, message: Fitur belum aktif untuk akun ini., request_id: req_opaque}
    NotFound:
      description: Resource is missing or not visible to the current owner.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    PayloadTooLarge:
      description: JSON request body exceeds the 64 KiB API limit.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: {code: PAYLOAD_TOO_LARGE, message: Body JSON melebihi batas 64 KiB., request_id: req_opaque}
    RateLimited:
      description: The authenticated account exceeded the mutation limit for this operation group.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
        Retry-After:
          description: Whole seconds until the current rate-limit window resets.
          schema: {type: integer, minimum: 1}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: {code: RATE_LIMITED, message: Terlalu banyak permintaan. Coba lagi nanti., request_id: req_opaque}
    ServiceUnavailable:
      description: A required local or upstream dependency is unavailable.
      headers:
        Cache-Control:
          $ref: "#/components/headers/PrivateNoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string, pattern: "^req_"}
...
    RuleCondition:
      type: object
      additionalProperties: false
      required: [left, op, right]
      properties:
        left: {type: string, enum: [PRICE, EMA9, EMA20, RSI14]}
        op: {type: string, enum: ["<", "<=", ">", ">="]}
        right:
          oneOf:
            - {type: string}
            - {type: number}
    UserRuleDefinition:
      type: object
      additionalProperties: false
      required: [name, logic, signal_type, cooldown_sec, conditions]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        logic: {type: string, enum: [AND]}
        signal_type: {type: string, enum: [BUY]}
        cooldown_sec: {type: integer, minimum: 0, maximum: 86400}
        conditions:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: "#/components/schemas/RuleCondition"
    UserRuleCreateRequest:
      type: object
      additionalProperties: false
      required: [definition]
      properties:
        definition:
          $ref: "#/components/schemas/UserRuleDefinition"
    UserRuleUpdateRequest:
      type: object
      additionalProperties: false
      required: [version, definition]
      properties:
        version: {type: integer, minimum: 1}
        definition:
          $ref: "#/components/schemas/UserRuleDefinition"
    UserRuleDeleteRequest:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version: {type: integer, minimum: 1}
    SystemRuleResource:
      type: object
      required: [id, name, owner_type, read_only, definition_hash, schema_version, engine_version, version]
      properties:
        id: {type: string, enum: [default-scalping-v1]}
        name: {type: string}
        owner_type: {type: string, enum: [system]}
        read_only: {type: boolean, enum: [true]}
        definition_hash: {type: string, pattern: "^sha256:"}
        schema_version: {type: string}
        engine_version: {type: string}
        version: {type: integer, enum: [1]}
    UserRuleResource:
      type: object
      required: [id, name, owner_type, read_only, definition, definition_hash, schema_version, engine_version, version, created_at, updated_at]
      properties:
        id: {type: string, pattern: "^rule_[a-f0-9]{24}$"}
        name: {type: string, minLength: 1, maxLength: 100}
        owner_type: {type: string, enum: [user]}
        read_only: {type: boolean, enum: [false]}
        definition:
          $ref: "#/components/schemas/UserRuleDefinition"
        definition_hash: {type: string, pattern: "^sha256:"}
        schema_version: {type: string}
        engine_version: {type: string}
        version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    RuleResource:
      oneOf:
        - $ref: "#/components/schemas/SystemRuleResource"
        - $ref: "#/components/schemas/UserRuleResource"
    RuleListResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RuleResource"
        next_cursor: {type: string, nullable: true}
...